Creating a new game is now done in /newgame
When you press the play button in the /newgame page, it will send an axios post to /createNewGame
with fields:
{
selectedCO: selectedCO,
selectedCountry: selectedCountry,
selectedMap: selectedMap,
};
A new gameDocument will be inserted into startGame collection where the orangeStar.CO will be
selectedCO from above. The blueMoon fields are now empty strings except for username where it is
"..." There is also now a startDate attribute in each game.

Game entries in /startgames will have the blueMoon character be all black and have the username "..."

The /newgame page, /startgames page, and /currentgames page all checks if the user is logged in. If not, it will redirect them to /login. The /newgame page authentication is done with /authenticateUser where it only uses the loggedIn middleware.

Next steps:
SelectCharacter.js, modal.sass, and util.js will be for further implementation of the feature where a user joins another game. In the future, when a user presses on a pending game, the website will prompt them to select a character.

+++ server/server.js
@@ -85,7 +85,6 @@ app.post("/loginUser", passport.authenticate("local"), (req, res) => {
   req.session.username = req.user.username;
   req.session._id = req.user._id;
 
-
   //we send 200 because that means we logged in correctly
   res.sendStatus(200);
 });
@@ -208,6 +207,11 @@ function loggedIn(req, res, next) {
   }
 }
 
+app.get("/authenticateUser", loggedIn, (req, res, next) => {
+  console.log("user is authenticated looool");
+  res.sendStatus(200);
+});
+
 app.get("/logout", function (req, res, next) {
   req.logout(function (err) {
     if (err) {
@@ -264,16 +268,20 @@ app.get("/userInfo", (req, res) => {
 });
 
 app.post("/joinGame", loggedIn, async (req, res) => {
-  
   // getting database
   let dbConnect = database.getDatabase();
   // getting startGame colleciton
   let startGameCollection = dbConnect.collection("startGame");
   // retrieving the game
-  let myDoc = await startGameCollection.findOne({ _id: new mongo.ObjectId(req.body.index) });
+  let myDoc = await startGameCollection.findOne({
+    _id: new mongo.ObjectId(req.body.index),
+  });
   // Checking if player is already in the game
-  if (myDoc.playerState.orangeStar.username == req.session.username) {
-    return
+  if (
+    myDoc.playerState.orangeStar.username == req.session.username ||
+    myDoc.playerState.blueMoon.username == req.session.username
+  ) {
+    return;
   }
   // deleting the game
   let result = await startGameCollection.deleteOne({
@@ -287,7 +295,7 @@ app.post("/joinGame", loggedIn, async (req, res) => {
   // moving the game from startGame to currentGame
   let currentGameCollection = dbConnect.collection("currentGame");
   // setting the id of the new currentGame
-  
+
   myDoc.playerState.blueMoon._id = req.session._id;
   try {
     await currentGameCollection.insertOne(myDoc);
